Allow references to other top-level fields
The OpenAPI spec allows parameters and responses to be defined at the
top level and referenced elsewhere. The expand references implementation
was limited to only the "definitions" section (that is all anyone was
using - until now..)

Expand references to any top-level field. The field can not have a
forward slash, and the name must have at lease one character.